+++ CHANGELOG.md
@@ -21,6 +21,8 @@
   stream in all cases
 * [HexaPDF::Type::AcroForm::TextField] to validate the existence of a /MaxLen
   value for comb text fields
+* [HexaPDF::Type::AcroForm::TextField#field_value=] to check for the existence
+  of /MaxLen when setting a value for a comb text field
 
 
 ## 0.22.0 - 2022-03-26